+++ src/__tests__/posthog-surveys.test.ts
@@ -114,6 +114,25 @@ describe('posthog-surveys', () => {
             delete assignableWindow.__PosthogExtensions__
         })
 
+        describe('canRenderSurvey', () => {
+            it('should return false if surveys are not loaded', () => {
+                const result = surveys.canRenderSurvey(survey.id)
+                expect(result.visible).toBeFalsy()
+                expect(result.disabledReason).toEqual('SDK is not enabled or survey functionality is not yet loaded')
+            })
+
+            it('should return visible: true if surveys are loaded and the survey is eligible', () => {
+                mockPostHog.get_property.mockReturnValue([survey])
+                surveys['_surveyManager'] = new SurveyManager(mockPostHog as PostHog)
+                decideResponse.featureFlags[survey.targeting_flag_key] = true
+                decideResponse.featureFlags[survey.internal_targeting_flag_key] = true
+                decideResponse.featureFlags[survey.linked_flag_key] = true
+                const result = surveys.canRenderSurvey(survey.id)
+                expect(result.visible).toBeTruthy()
+                expect(result.disabledReason).toBeUndefined()
+            })
+        })
+
         describe('checkSurveyEligibility', () => {
             beforeEach(() => {
                 // mock getSurveys response